Create DB module and update some comments
